[#72] 지도 페이지 json error 수정
