+++ src/app/(header)/search/page.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/components/index';
 import usePagination from '@/hooks/usePagination';
 import axios from 'axios';
-import { Suspense, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import kakaoMarkerGenerator, {
   CampPlaceType,
 } from '../../_utils/kakaoMarkerGenerator';
@@ -56,27 +56,16 @@ function Page({ searchParams }: SearchParamsType) {
   };
 
   useEffect(() => {
-    const { location, checkIn, checkOut, group: groupParam } = searchParams;
-
-    let group = { adult: 0, child: 0, pet: 0 };
-    if (groupParam) {
-      try {
-        group = JSON.parse(groupParam);
-      } catch (e) {
-        console.error('Error parsing group params:', e);
-      }
-    }
-
-    const totalNumberOfPeople = group.adult + group.child;
+    const { location, checkIn, checkOut, group: groupTest } = searchParams;
 
     const fetch = async () => {
       const response = await axios.get<DataType>(
         `data/mapPositionsMockData.json`,
       );
-      const { result } = response.data;
 
+      const { result } = response.data;
       const filteredResult =
-        location && checkIn && checkOut && totalNumberOfPeople > 0
+        location && checkIn && checkOut
           ? result.filter((item) => item.address.includes(location))
           : result;
 
@@ -100,9 +89,7 @@ function Page({ searchParams }: SearchParamsType) {
       />
       <div className='border-bg-gray200 relative z-[99] border-b bg-white px-40pxr pb-28pxr pt-20pxr mobile:flex mobile:p-16pxr'>
         <div className='m-auto w-full max-w-1360pxr'>
-          <Suspense>
-            <SearchBarForSearch />
-          </Suspense>
+          <SearchBarForSearch searchParams={searchParams} />
         </div>
         <div className='mobile:flex-center z-[99] flex gap-12pxr tabletMin:w-full'>
           <SearchFilter />
